+++ packages/widget_toolkit/lib/search_picker.dart
@@ -5,3 +5,4 @@ export 'src/lib_search_picker/services/search_picker_service.dart';
 export 'src/lib_search_picker/theme/search_picker_theme.dart';
 export 'src/lib_search_picker/ui_components/show_search_picker_bottom_sheet.dart';
 export 'src/lib_search_picker/views/search_picker_page.dart';
+export 'src/lib_search_picker/views/search_picker_page_with_dependencies.dart';